WIB-14 Rename | models 디렉토리 auth 추가, 기존 로그인 관련 모듈을 이동하였다.
