--- client/models/ValidateEmail.ts
@@ -1,4 +1,0 @@
-export const ValidateEmail = (email: string): boolean => {
-  const regex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
-  return regex.test(email);
-};
