--- client/models/ValidatePassword.ts
@@ -1,5 +1,0 @@
-export const ValidatePassword = (pw: string): boolean => {
-  const regex =
-    /^(?=.*?[A-Z])(?=.*?[a-z])(?=.*?[0-9])(?=.*?[#?!@$ %^&*-]).{6,18}$/;
-  return regex.test(pw);
-};
